click handler to keep the text displayed
Started to try and figure out the time blocks color - but decided to move onto the saving calendar events. I created click handler on the save buttons so they save and continue to display the value once typed - even when the page is refreshed. Only working on the 9am hour atm.