Local storage get and set DONE
Played around with different ways to set and get the local storage array information. I ended up creating a saveButton that on a click event triggers the previous sibling (the input value) to be saved into an array at the buttons index number. I get the data attribute of each button to equal an index number. Then in a for loop it retreives the array values and displays into the html inputs using dom traveral selectors. Next I will work of the hour colours.